+++ packages/riverpod_cli/lib/src/migrate/imports.dart
@@ -1,5 +1,6 @@
 // ignore: deprecated_member_use
-import 'package:analyzer/analyzer.dart';
+import 'package:analyzer/dart/ast/ast.dart';
+import 'package:analyzer/dart/ast/visitor.dart';
 import 'package:codemod/codemod.dart';
 
 /// A suggestor that yields changes to deprecated import directives